requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
pandas==1.3.1
Flask==2.0.2
requests==2.26.0
matplotlib==3.4.2
torchvision==0.11.1+cpu
Pillow==8.4.0
torch==1.10.0+cpu
gunicorn==20.0.4
